created mapper class, added create entity method in Mapper and connected it to serrvice class

+++ src/main/java/com/example/springlabb/Movies/MovieRepository.java
@@ -2,5 +2,5 @@ package com.example.springlabb.Movies;
 
 import org.springframework.data.repository.ListCrudRepository;
 
-public interface MovieRepository extends ListCrudRepository<Movie, Integer> {
+public interface MovieRepository extends ListCrudRepository<Movie, Long> {
 }
